refactor(xorq_datafusion)!: flatten DataFusionBackend inheritance
Merge two-level xorq_datafusion → datafusion.Backend hierarchy into a
single Backend class. Deletes the datafusion/ subdirectory entirely.

Fixes bundled with the merge:
- register(): match/case dispatch, single deregister_table, one return;
  add unbound-expression guard; ir.Table/ir.Expr arms replace verbose
  ibis.expr.types.* references; drop redundant hasattr guards
- _compile_pyarrow_udaf: param_name loop var shadowed outer name;
  materialize parameters to list, guard empty case
- _to_pyarrow_batches / to_parquet: remove redundant _import_pyarrow()
  calls — pyarrow unconditionally imported at module level
- register_table_provider: narrow source to ir.Table; guard None name
- translate_sort: move to Backend._translate_sort as @staticmethod
- Promote pandas/pyarrow.dataset/pyarrow.parquet to module-level imports

BREAKING CHANGE: truncate_table removed — DataFusion does not implement
DELETE DML; the method was a silent no-op.

BREAKING CHANGE: read_postgres removed — database param was silently
ignored by _from_url kwargs.update(); no callers found in codebase.
