+++ python/xorq/backends/xorq_datafusion/tests/test_provider.py
@@ -2,7 +2,7 @@ import pyarrow as pa
 import pytest
 
 import xorq.api as xo
-from xorq.backends.xorq_datafusion.datafusion.provider import IbisTableProvider
+from xorq.backends.xorq_datafusion.provider import IbisTableProvider
 
 
 @pytest.fixture
